fix(frontend): tighten session operation handling
